requirements.txt
Package                  Version
------------------------ -----------
asgiref                  3.8.1
attrs                    24.2.0
certifi                  2024.8.30
cffi                     1.17.1
charset-normalizer       3.3.2
Django                   5.1.1
django-cors-headers      4.4.0
django-filter            24.3
django-grappelli         4.0.1
django-phonenumber-field 8.0.0
djangorestframework      3.15.2
h11                      0.14.0
idna                     3.10
outcome                  1.3.0.post0
phonenumbers             8.13.47
pillow                   10.4.0
pip                      24.2
psycopg                  3.2.3
pycparser                2.22
Pygments                 2.18.0
PySocks                  1.7.1
requests                 2.32.3
selenium                 4.25.0
sniffio                  1.3.1
sortedcontainers         2.4.0
sqlparse                 0.5.1
trio                     0.26.2
trio-websocket           0.11.1
typing_extensions        4.12.2
tzdata                   2024.2
urllib3                  2.2.3
websocket-client         1.8.0
wsproto                  1.2.0
